move logic to model from controller
